Fix examples
The examples did not compile since cmdInit was changed to
accept a Stream reference instead of a serial port speed. This
change fixes all examples. It also adds a new function
cmdGetStream(). This allows command functions to get access to
the same stream that is used by the command interpreter itself.
If you change a sketch from using the serial port to using a
different stream, you can simply change the initialisation of the
cmd module and all command functions can remain unchanged.

The commit also contains a fix to ignore \n characters as they
are usually not part of a valid command. This way, the code works
with all terminals, no matter if they send \r or \r\n line endings.